fix to phones and mod functionalities

+++ src/main/java/com/data/safehaven/dtos/AdministradorDto.java
@@ -11,7 +11,7 @@ public record AdministradorDto(Long id,
                                String correoElectronico,
                                String password,
                                Integer edad,
-                               Integer telefono,
+                               Long telefono,
                                String sexo,
                                Date fechaDeNacimiento,
                                String cargo) {
